% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_df_de_atrito.R
\name{cria_df_de_atrito}
\alias{cria_df_de_atrito}
\title{Create an attrition table for a panel file}
\usage{
cria_df_de_atrito(data, panel)
}
\arguments{
\item{data}{The input data frame, preferably a PNADc panel file with all 5
interviews of the individuals of that panel, with a certain panel method
already applied.}

\item{panel}{The identification strategy whose friction will be identified. Can either be "basic", "advanced" or "household"}
}
\value{
A data frame summarizing missing interviews and the percentage of
interviews attended for the individuals forming that panel.
}
\description{
This function generates a summary dataframe indicating the count of missing
interviews for each individual and the percentage of interviews attended,
calculating the attrition for a PNADc panel.
}
\examples{
\dontrun{
data <- read.csv("path/to/your/database/pnadc_panel_6.csv")
attrition_panel_6_PNADc <- cria_df_de_atrito(data, panel = "advanced")
}

}
\keyword{internal}
