% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_pnad.R
\encoding{UTF-8}
\name{datazoom_pnadc}
\alias{datazoom_pnadc}
\title{Carregando dados brutos}
\usage{
datazoom_pnadc(diretorio_dados, ...)
}
\arguments{
\item{diretorio_dados}{Diretório onde os microdados originais em formato de texto estão armazenados}

\item{...}{vetores com datas das pesquisas de interesse no  formato \code{c('trimestre', 'ano')}}
}
\value{
Lista de dataframes, sendo cada entrada um trimestre/ano em \code{...}
}
\description{
Carregando dados brutos
}
\examples{
datazoom_pnadc("./Desktop", "./Desktop", c(1, 2000), c(2, 2000))
}
