% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{UTF-8}
\name{dictionary_pnadcovid_eng}
\alias{dictionary_pnadcovid_eng}
\title{Dictionary for PNAD-Covid 19 microdata - English}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 594 rows and 5 columns.
}
\usage{
dictionary_pnadcovid_eng
}
\description{
This dataframe contains information on PNAD-Covid 19 data. It describes,
in Portuguese each variable and the possible values assumed by them .
All information is in English. Check \code{dictionary_pnadcovid_pt.br}
for the version in Portuguese
}
\seealso{
\code{\link{dictionary_pnadcovid_pt.br}}
}
\keyword{datasets}
