% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_quarter}
\alias{download_quarter}
\title{Download microdata from a selected year and quarter}
\usage{
download_quarter(quarter, year, dir = tempdir())
}
\arguments{
\item{quarter}{Number between 1 and 4 indicating the quarter}

\item{year}{Number between 2012 and present year}

\item{dir}{Directory where data should be saved}
}
\value{
filename of the downloaded file
}
\description{
Download microdata from a selected year and quarter
}
