% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrating_functions.R
\encoding{UTF-8}
\name{load_pnadcontinua}
\alias{load_pnadcontinua}
\title{Loads and cleans PNAD Continua microdata}
\usage{
load_pnadcontinua(language = "eng", sources, download_directory = getwd())
}
\arguments{
\item{language}{Choose whether the data should come in Portuguese or in English. Default is
\code{language = 'eng'}. For portuguese, write instead \code{language = 'pt'}.}

\item{sources}{A number of different sources are supported:

Passing a list of dates will download the corresponding data from IBGE's website.

Passing a string with a directory's path will read data from all files named
"\code{path}/PNADC_XXXX.txt".

Alternatively, \code{sources} may be a list of full file paths}

\item{download_directory}{In case \code{sources} is such that data is downloaded
from IBGE, where should it be stored? Default is the working directory}
}
\value{
A list of dataframes with the microdata from each
required period. Downloaded raw data are stored in a folder called "PNADC_microdata"
on the directory specified in argument \code{download_directory}
}
\description{
Loads and cleans PNAD Continua microdata
}
\examples{

\dontrun{
To download data, set \code{sources} as a list of vectors
of time periods

dates <- list(c(1, 2012), c(2, 2012))

microdata <- load_pnadcontinua(language = 'eng',
                        sources = dates,
                        download_directory = './Desktop')

To load the data from a folder:

microdata <- load_pnadcontinua(language = 'eng',
                        sources = './Desktop/folder_name')

To load an individual .txt file corresponding to a given period of the survey:

  microdata <- load_pnadcontinua(sources = './PNADC_012020.txt')
}

}
