% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_pnad.R
\encoding{UTF-8}
\name{pnadc_painel_basico}
\alias{pnadc_painel_basico}
\title{Painel básico}
\usage{
pnadc_painel_basico(build_data = TRUE, ...)
}
\arguments{
\item{build_data}{Default \code{TRUE}.
Se \code{TRUE}, implementa primeiro \code{\link{datazoom_pnadc}} e depois
monta paineis de indiv??duos. Se \code{FALSE}, a função constrói paneis a partir de dados já carregados no R}

\item{dados_prontos}{Bases de dados para diferentes trimestres da PNAD cont??nua.
Necessário se \code{build_data = FALSE}}

\item{local_dados}{Diretório onde os microdados originais em formato de texto estão armazenados
caso \code{build_data = TRUE}}

\item{periodos}{Lista de vetores com per??odos de interesse no formato
\code{periodos = list(c(trimestre1, ano1), c(trimestre2, ano2), ...)}}
}
\value{
Lista de dataframes, sendo cada entrada um trimestre/ano
}
\description{
Painel básico
}
\examples{
PNADC_2012 <- datazoom_pnadc(
  diretorio_dados = "./Desktop",
  c(1, 2012), c(2, 2012)
)

teste <- pnadc_painel_basico(
  build_data = FALSE,
  dados_prontos = PNADC_2012
)

teste2 <- pnadc_painel_basico(
  build_data = TRUE,
  local_dados = "./pnadcontinua",
  periodos = list(c(1, 2012), c(2, 2012))
)
}
