% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pnadc_panel.R
\name{build_pnadc_panel}
\alias{build_pnadc_panel}
\title{Build PNADc Panel}
\usage{
build_pnadc_panel(dat, panel)
}
\arguments{
\item{dat}{Data frame with PNADC data, sorted into a single panel.}

\item{panel}{A \code{character} with the type of panel identification. Use "none" for no paneling, "basic" for basic paneling, and "advanced" for advanced paneling.}
}
\value{
A modified dataset with added identifiers for household (\code{id_dom}) and individual (\code{id_ind} or \code{id_rs}) based on the chosen panel algorithm.
}
\description{
This function builds a panel dataset from PNADC data, identifying households and individuals
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example usage:

panel_data <- build_pnadc_panel(dat = pnad_sample, panel = "basic")
\dontshow{\}) # examplesIf}
}
