% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pnadc_panel.R
\name{build_pnadc_panel}
\alias{build_pnadc_panel}
\title{Build PNADc Panel}
\usage{
build_pnadc_panel(dat, panel)
}
\arguments{
\item{dat}{The current PNAD observations.}

\item{panel}{The type of panelling transformation you wish to apply to \code{dat}. Use "none" for no paneling, "basic" for basic paneling, and "advanced" for advanced paneling.}
}
\value{
A modified dataset with added identifiers for household (\code{id_dom}) and individual (\code{id_ind}) based on the chosen panel algorithm.
}
\description{
This function is designed to build a panel dataset from PNADC data based on the chosen panel algorithm. It performs basic and, if specified, advanced identification steps to create household and individual identifiers for panel construction.
}
\examples{
\dontrun{
# Example usage:
data <- read.csv("path/to/PNADC_data.csv")
panel_data <- build_pnadc_panel(dat = data, panel = "basic")
}

}
\references{
Ribas, Rafael Perez, and Sergei Suarez Dillon Soares. Sobre o painel da Pesquisa Mensal de Emprego (PME) do IBGE. No. 1348. Texto para discussão, 2008.
Data Zoom (2023). Data Zoom: Simplifying Access To Brazilian Microdata.
}
\author{
Data Zoom
}
