% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_panel_attrition.R
\name{calculate_panel_attrition}
\alias{calculate_panel_attrition}
\title{Create an attrition table for a PNADc panel file}
\usage{
calculate_panel_attrition(data, panel)
}
\arguments{
\item{data}{Input data frame (PNADc panel).}

\item{panel}{ID strategy: "basic" (id_ind), "advanced" (id_rs), or "households" (id_dom).}
}
\value{
A data frame with 8 columns detailing unconditional and conditional attrition.
}
\description{
Create an attrition table for a PNADc panel file
}
