% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_df_de_atrito.R
\name{cria_df_de_atrito}
\alias{cria_df_de_atrito}
\title{Create an attrition table for a panel file}
\usage{
cria_df_de_atrito(data, panel)
}
\arguments{
\item{data}{The input data frame, preferably a PNADc panel file.}

\item{panel}{The identification strategy: "basic", "advanced_1", "advanced_2", "advanced_3" or "households".}
}
\value{
A data frame summarizing missing interviews and the tracking rates.
}
\description{
This function generates a summary dataframe indicating the count of missing
interviews for each individual and the unconditional tracking rates.
}
