% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pnadc.R
\name{load_pnadc}
\alias{load_pnadc}
\title{Load PNADc Data}
\usage{
load_pnadc(
  save_to = getwd(),
  year,
  quarter = 1:4,
  panel = "advanced",
  raw_data = FALSE,
  language = "pt"
)
}
\arguments{
\item{save_to}{The directory in which the user desires to save the downloaded files.}

\item{year}{The years of the PNADc the user would like to download.}

\item{quarter}{The quarters within those years to be downloaded.}

\item{panel}{Which panel algorithm to apply to this data (none, basic, or advanced). Check the README for a detailed explanation.}

\item{raw_data}{A command to define if the user would like to download the raw or treated data.}
}
\value{
A message indicating the successful save of panel files.
}
\description{
This function downloads PNADc data for specified years and quarters, applies panel identification algorithms, and saves the data to .rds (the individual quarter files) and .csv (the individual panel files) files for each panel.
}
\examples{
\dontrun{
load_pnadc(
  save_to = "Directory/You/Would/like/to/save/the/files",
  year = 2016,
  quarter = 1:4,
  panel = "basic",
  raw_data = FALSE
)
}
}
