% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pnadc.R
\name{load_pnadc}
\alias{load_pnadc}
\title{Load Continuous PNAD Data}
\usage{
load_pnadc(
  save_to = getwd(),
  years,
  quarters = 1:4,
  panel = "advanced",
  raw_data = FALSE
)
}
\arguments{
\item{save_to}{A \code{character} with the directory in which to save the downloaded files.}

\item{years}{A \code{numeric} indicating for which years the data will be loaded, in the format YYYY. Can be any vector of numbers, such as 2010:2012.}

\item{quarters}{The quarters within those years to be downloaded. Can be a numeric vector or a list of vectors, for different quarters per year.}

\item{panel}{A \code{character} choosing the panel algorithm to apply ("none", "basic", or "advanced"). For details, check \code{vignette("BUILD_PNADC_PANEL")}}

\item{raw_data}{A \code{logical} setting the return of raw (\code{TRUE}) or processed (\code{FALSE}) variables.}
}
\value{
A message indicating the successful save of panel files.
}
\description{
This function downloads PNADC data and applies panel identification algorithms
}
\examples{
\dontrun{
load_pnadc(
  save_to = "Directory/You/Would/like/to/save/the/files",
  years = 2016,
  quarters = 1:4,
  panel = "basic",
  raw_data = FALSE
)
}
}
