% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrating_functions.R
\encoding{UTF-8}
\name{load_pnadc}
\alias{load_pnadc}
\title{Loads and cleans PNAD Continua microdata from a specified directory.}
\usage{
load_pnadc(lang = "english", sources, download_directory = getwd())
}
\arguments{
\item{lang}{Should the data come in Portuguese or in English. Default is
\code{lang = 'english'}}

\item{sources}{A number of different sources are supported:

Passing a list of dates will download the corresponding data from IBGE's website.

Passing a string with a directory's path will read data from all files named
"\code{path}/PNADC_XXXX.txt".

Alternatively, \code{sources} may be a list of full file paths}

\item{download_directory}{In case \code{sources} is such that data is downloaded
from IBGE, where should it be stored? Default is the working directory}
}
\value{
A list of dataframes with the microdata from each
required period.
}
\description{
Loads and cleans PNAD Continua microdata from a specified directory.
}
\examples{

\dontrun{
To download data, set \code{sources} as a list of vectors
of time periods

dates <- list(c(1, 2012), c(2, 2012))

microdata <- load_pnadc(lang = 'english',
                        sources = dates,
                        download_directory = './Desktop')

To load the data from a folder:

microdata <- load_pnadc(lang = 'english',
                        sources = './Desktop/folder_name')

To load an individual .txt file corresponding to a given period of the survey:

  microdata <- load_pnadc(sources = './PNADC_012020.txt')
}

}
